Fix 'expression is not assignable'
When trying to compile the project I hit the following issue:

```
rbtree.c:1086:24: error: expression is not assignable
    RHASH_IFNONE(hash) = IFNONE(self);
```

The problem is that this API is no longer public. It was made private in:
https://bugs.ruby-lang.org/issues/9889

With this change we no longer support smooth transition between RBTree and Hash,
which preserves default value. I'm not 100% sure that dropping support is the
right way but there's only much I can do. My Ruby C API knowledge is not very
deep.

This can potentially break someone's code but it doesn't seem to be too
critical.