The taint mechanism was deprecated in Ruby 2.7
The Ruby core team decided to deprecate the taint mechanism in Ruby 2.7
and will remove that in Ruby 3.

https://bugs.ruby-lang.org/issues/16131
ruby/ruby#2476

In Ruby 2.7, `Object#{taint,untaint,trust,untrust}` and related
functions in the C-API no longer have an effect (all objects are always
considered untainted), and are now warned with deprecation messages.

The format of Proc inspect has changed too. Adjusted the affected test.

[Fixes #6]