Fixes for: #16 Error installing on M1, Ruby 3.x (#17)
For: https://github.com/kyrylo/rbtree3/issues/16

The change is to explicitly cast `key1` and `key2` as `VALUE` before calling the `TYPE` macro, as well as casting them as `VALUE` before passing them as arguments to `rb_str_cmp()` and `rb_funcall()`. This ensures that the correct `VALUE` type is used.

The `arg->key` argument is of type `const void*`, which is not compatible with `VALUE` (aka unsigned long) expected by the `TYPE` macro. To fix this error, `arg->key` needs to be casted to `VALUE` before passing it to `TYPE`.

Forked this to try and see if I can fix it. This is the build error that's happening on Apple M1 computers: `https://github.com/kyrylo/rbtree3/issues/16`

Locally, I was able to install this gem on my Apple M1 with Ruby 3.0.6 with these changes.

I'm not a C programmer, so please scrutinize.